requirements.txt
allure-pytest==2.8.40
pytest==6.2.3
pytest-xdist==2.2.1
selenium==3.141.0
webdriver-manager==3.3.0
